requirements.txt
langchain>=0.3.0
langchain_openai>=0.2.0
langchain-core>=0.3.0
langsmith>=0.3.0
langchain-experimental>=0.3.0
langserve>=0.1.0
langchainhub>=0.1.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.5.1
openai>=1.13.3
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.100.0
serpapi>=0.1.0
lxml>=4.9.3
streamlit>=1.32.0
markdown>=3.4.3
matplotlib>=3.10.0
altair>=5.5.0
pandas>=2.0.0
iointel>=0.1.0
asyncio 
langchain
langchain-google-genai
langchain-openai
langchain-community
chromadb
transformers
torch
sentence-transformers
python-dotenv
watchdog
pypdf
rouge_score
datasets
openai
# RAGAS dependencies
ragas>=0.1.0
tqdm>=4.66.0